feat: 翻訳者クラスとJavaだけの翻訳者クラスを分離
